refactor: move case logic into context
